pyproject.toml
[project]
name = "renamr"
version = "0.1.0"
description = "CLI tool that renames local files using AI-powered metadata extraction"
authors = [
    { name = "Silas Pignotti", email = "pignottisilas@gmail.com" }
]
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
dependencies = [
    "litellm>=1.82.1",
    "pillow>=12.1.1",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "pymupdf>=1.27.2",
    "pypdf>=6.8.0",
    "structlog>=25.5.0",
    "typer>=0.24.1",
]

[project.scripts]
renamr = "renamr.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "nox>=2026.2.9",
    "pip-audit>=2.9.0",
    "pyright>=1.1.408",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "ruff>=0.15.5",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "RUF"]

[tool.pyright]
include = ["src", "tests"]
typeCheckingMode = "standard"
pythonVersion = "3.12"
venvPath = "."
venv = ".venv"

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
